requirements.txt
flask==2.3.2
flask-jwt-extended==4.5.2
pymongo==4.4.1
python-dotenv==1.0.0
pyjwt==2.8.0
cryptography==41.0.3